refactor: move CPU instensive tasks into worker pool
