+++ src/server/package.json
@@ -64,6 +64,7 @@
     "reflect-metadata": "0.1.13",
     "rollbar": "2.19.2",
     "rxjs": "7.4.0",
+    "threads": "1.7.0",
     "topojson-client": "3.1.0",
     "typeorm": "0.2.41"
   },
