+++ src/server/src/projects/controllers/projects.controller.ts
@@ -41,7 +41,6 @@ import {
   DistrictsDefinition,
   ProjectId,
   PublicUserProperties,
-  GeoUnitHierarchy,
   UserId,
   ProjectTemplateId,
   IUser,
@@ -220,7 +219,7 @@ export class ProjectsController implements CrudController<Project> {
   ) {
     // Districts definition is optional. Use it if supplied, otherwise use all-unassigned.
     const districtsDefinition =
-      dto.districtsDefinition || new Array(geoCollection.hierarchy.length).fill(0);
+      dto.districtsDefinition || new Array(geoCollection.districtsDefLength).fill(0);
     const lockedDistricts = new Array(dto.numberOfDistricts).fill(false);
     const numberOfMembers = dto.numberOfMembers || new Array(dto.numberOfDistricts).fill(1);
     return {
@@ -353,18 +352,6 @@ export class ProjectsController implements CrudController<Project> {
 
   // Helper for obtaining a topology for a given S3 URI, throws exception if not found
   async getGeoUnitTopology(s3URI: string): Promise<GeoUnitTopology> {
-    const geoCollection = await this.getGeoUnitProperties(s3URI);
-    if (!("topology" in geoCollection)) {
-      throw new NotFoundException(
-        `Topology ${s3URI} is archived`,
-        MakeDistrictsErrors.TOPOLOGY_NOT_FOUND
-      );
-    }
-    return geoCollection;
-  }
-
-  // Helper for obtaining a topology or props for a given S3 URI, throws exception if not found
-  async getGeoUnitProperties(s3URI: string): Promise<GeoUnitTopology> {
     const geoCollection = await this.topologyService.get(s3URI);
     if (!geoCollection) {
       throw new NotFoundException(
@@ -389,7 +376,7 @@ export class ProjectsController implements CrudController<Project> {
     readonly regionConfig: IRegionConfig;
   }): Promise<DistrictsGeoJSON> {
     const geoCollection = await this.getGeoUnitTopology(regionConfig.s3URI);
-    const geojson = geoCollection.merge({
+    const geojson = await geoCollection.merge({
       districtsDefinition,
       numberOfDistricts,
       user,
@@ -469,50 +456,6 @@ export class ProjectsController implements CrudController<Project> {
     await convert(formattedGeojson, response, { layer: "districts" });
   }
 
-  exportToCsv(
-    geoCollection: GeoUnitTopology,
-    districtsDefinition: DistrictsDefinition
-  ): [string, number][] {
-    const baseGeoLevel = geoCollection.definition.groups.slice().reverse()[0];
-    const baseGeoUnitProperties = geoCollection.topologyProperties[baseGeoLevel];
-
-    // First column is the base geounit id, second column is the district id
-    const mutableCsvRows: [string, number][] = [];
-
-    // The geounit hierarchy and district definition have the same structure (except the
-    // hierarchy always goes out to the base geounit level). Walk them both at the same time
-    // and collect the information needed for the CSV (base geounit id and district id).
-    const accumulateCsvRows = (
-      defnSubset: number | GeoUnitHierarchy,
-      hierarchySubset: GeoUnitHierarchy
-    ) => {
-      hierarchySubset.forEach((hierarchyNumOrArray, idx) => {
-        const districtOrArray = typeof defnSubset === "number" ? defnSubset : defnSubset[idx];
-        if (typeof districtOrArray === "number" && typeof hierarchyNumOrArray === "number") {
-          // The numbers found in the hierarchy are the base geounit indices of the topology.
-          // Access this item in the topology to find it's base geounit id.
-          const props: any = baseGeoUnitProperties[hierarchyNumOrArray];
-          const baseId = props[baseGeoLevel] as string;
-          mutableCsvRows.push([baseId, districtOrArray]);
-        } else if (typeof hierarchyNumOrArray !== "number") {
-          // Keep recursing into the hierarchy until we reach the end
-          accumulateCsvRows(districtOrArray, hierarchyNumOrArray);
-        } else {
-          // This shouldn't ever happen, and would suggest a district definition/hierarchy mismatch
-          this.logger.error(
-            "Hierarchy and districts definition mismatch",
-            districtOrArray.toString(),
-            hierarchyNumOrArray.toString()
-          );
-          throw new InternalServerErrorException();
-        }
-      });
-    };
-    accumulateCsvRows(districtsDefinition, geoCollection.hierarchyDefinition);
-
-    return mutableCsvRows;
-  }
-
   @UseInterceptors(CrudRequestInterceptor)
   @UseGuards(OptionalJwtAuthGuard)
   @Get(":id/export/csv")
@@ -522,9 +465,9 @@ export class ProjectsController implements CrudController<Project> {
     @Param("id") projectId: ProjectId
   ): Promise<string> {
     const project = await this.getProject(req, projectId);
-    const geoCollection = await this.getGeoUnitProperties(project.regionConfig.s3URI);
+    const geoCollection = await this.getGeoUnitTopology(project.regionConfig.s3URI);
     const baseGeoLevel = geoCollection.definition.groups.slice().reverse()[0];
-    const csvRows = this.exportToCsv(geoCollection, project.districtsDefinition);
+    const csvRows = await geoCollection.exportToCSV(project.districtsDefinition);
 
     return stringify(csvRows, {
       header: true,
@@ -636,7 +579,7 @@ export class ProjectsController implements CrudController<Project> {
     }
     validateNumberOfMembers(dto, existingProject.numberOfDistricts);
 
-    const staticMetadata = (await this.getGeoUnitProperties(existingProject.regionConfig.s3URI))
+    const staticMetadata = (await this.getGeoUnitTopology(existingProject.regionConfig.s3URI))
       .staticMetadata;
     const allowedDemographicFields = getDemographicsMetricFields(staticMetadata).map(
       ([, field]) => field
